core/permissions.py
from rest_framework.permissions import SAFE_METHODS, BasePermission


class IsStaffOrReadOnly(BasePermission):
    """Allow writes only to authenticated staff members."""

    def has_permission(self, request, view):  # pragma: no cover - simple
        if request.method in SAFE_METHODS:
            return True
        user = request.user
        return bool(user and user.is_authenticated and user.is_staff)
