requirements.txt
asgiref==3.10.0
charset-normalizer==3.4.4
Django==4.2.26
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
drf-spectacular==0.27.2
pandas==2.2.3
pillow==11.3.0
python-dotenv==1.2.1
reportlab==4.4.4
sqlparse==0.5.3
typing_extensions==4.15.0
